forgot to remove float cast for print statement

+++ catkin_ws_private/src/control/src/control_node.cpp
@@ -128,7 +128,7 @@ class Control {
         
         pubSteeringPose_.publish(steeringPose_msg);
         ROS_INFO("control_node setDirection: pos(%.2f/%.2f) dst(%.2f/%.2f) dist: %.2f speed: %d",
-                 pos.x,pos.y, dst.x, dst.y, distance, (float)speed_msg.data);
+                 pos.x,pos.y, dst.x, dst.y, distance, speed_msg.data);
 #endif
       }
